stackoverflow/requirements.txt
aiohttp==3.6.2
alembic==0.9.3.dev0
apturl==0.5.2
argcomplete==1.8.1
argh==0.26.2
asgiref==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Babel==2.4.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
Brotli==0.6.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
construct==2.8.16
cryptography==2.1.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.1.7
django-cors-headers==3.7.0
django-premailer==0.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
feedparser==5.2.1
Flask==1.1.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
greenlet==0.4.12
h2==3.0.1
hpack==3.0.0
html2text==2016.9.19
html5lib==0.999999999
httplib2==0.9.2
hyperframe==5.1.0
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsbeautifier==1.6.4
kaitaistruct==0.7
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
ldap3==2.4.1
louis==3.5.0
lxml==4.6.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==3.10.0
mitmproxy==2.0.2
msgpack==0.5.6
multidict==4.7.6
neovim==0.2.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
paramiko==2.0.0
passlib==1.7.1
pathtools==0.1.2
pexpect==4.2.1
Pillow==5.1.0
premailer==3.0.0
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.5
pyasn1==0.2.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.1.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
pyperclip==1.5.32
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.25.1
requests-unixsocket==0.1.5
ruamel.yaml==0.13.14
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
slackclient==2.7.2
sortedcontainers==1.5.7
speaklater==1.4
SQLAlchemy==1.1.11
sqlparse==0.4.1
ssh-import-id==5.7
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
tornado==4.5.3
typing-extensions==3.7.4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
urwid==1.3.1
usb-creator==0.3.3
wadllib==1.3.2
watchdog==0.8.3
webencodings==0.5
Werkzeug==1.0.1
WTForms==2.1
xkit==0.0.0
yarl==1.4.2
zope.interface==4.3.2
